# speed up the new factor graphs