Adds more specializations, and changes some particularly hot code to use pointers instead of collections.

Makes Pose2SLAM ~4x faster and Pose3SLAM ~3x faster.

Before:
```
swift build -c release -Xswiftc -cross-module-optimization && .build/release/Benchmarks --filter NewFactorGraph
name                                                                    time             std        iterations
--------------------------------------------------------------------------------------------------------------
Pose2SLAM: NewFactorGraph, Intel, 10 Gauss-Newton steps, 500 CGLS steps  1868228204.0 ns ±   0.00 %          1
Pose3SLAM: NewFactorGraph, sphere2500, 30 LM steps, 200 CGLS steps      19669232876.0 ns ±   0.00 %          1
```

After:
```
swift build -c release -Xswiftc -cross-module-optimization && .build/release/Benchmarks --filter NewFactorGraph
name                                                                    time            std        iterations
-------------------------------------------------------------------------------------------------------------
Pose2SLAM: NewFactorGraph, Intel, 10 Gauss-Newton steps, 500 CGLS steps  444424233.0 ns ±   0.00 %          1
Pose3SLAM: NewFactorGraph, sphere2500, 30 LM steps, 200 CGLS steps      6116728990.0 ns ±   0.00 %          1
```