@dabrahams 2020-06-18
Also, after using the new dispatch stuff it would be good to check to see if all of these hand-optimizations are still needed.  Sucks, right?